Fixed a small bug with the print frequency in application.

+++ source/pysph/solver/utils.py
@@ -244,15 +244,17 @@ def _mkdir(newdir):
         - parent directory(ies) does not exist, make them as well
     """
     if os.path.isdir(newdir):
-        status, op = commands.getstatusoutput("rm -r %s"%(newdir))
+        if not newdir == ".":
+            status, op = commands.getstatusoutput("rm -r %s"%(newdir))
 
-        if status != 0:
-            raise OSError("Could not delete contents of directory %s"%(newdir))
+            if status != 0:
+                err = "Could not delete contents of directory %s"%(newdir)
+                raise OSError(msg)
 
-        status, op = commands.getstatusoutput("mkdir  %s"%(newdir))
+            status, op = commands.getstatusoutput("mkdir  %s"%(newdir))
 
-        if status != 0:
-            raise OSError("Could not create directory %s"%(newdir))
+            if status != 0:
+                raise OSError("Could not create directory %s"%(newdir))
 
     elif os.path.isfile(newdir):
         raise OSError("a file with the same name as the desired " \
